+++ lib/src/core/core.dart
@@ -3,7 +3,7 @@ export '../modules/data_cache/domain/constants/cache_constants.dart';
 export 'config/cache_config.dart';
 export 'exceptions/auto_cache_manager_exception.dart';
 
-export 'services/directory_provider/directory_provider_service.dart';
+export 'services/directory_service/directory_provider_service.dart';
 export 'services/injector/injector.dart';
 
 export 'utils/either.dart';
