+++ lib/src/core/services/directory_service/exceptions/directory_provider_exceptions.dart
