+++ lib/src/modules/data_cache/external/datasources/query_data_cache_datasource.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import '../../../../core/services/cryptography/i_cryptography_service.dart';
+import '../../../../core/services/cryptography_service/i_cryptography_service.dart';
 import '../../../../core/services/storages/prefs/i_prefs_service.dart';
 import '../../domain/entities/cache_entity.dart';
 import '../../infra/datasources/i_query_data_cache_datasource.dart';
