+++ test/src/core/services/directory_provider/directory_provider_service_test.dart
@@ -1,8 +1,8 @@
 import 'dart:io';
 
 import 'package:auto_cache_manager/src/core/core.dart';
-import 'package:auto_cache_manager/src/core/services/directory_provider/exceptions/directory_provider_exceptions.dart';
-import 'package:auto_cache_manager/src/core/services/directory_provider/path_provider/path_provider_service.dart';
+import 'package:auto_cache_manager/src/core/services/directory_service/exceptions/directory_provider_exceptions.dart';
+import 'package:auto_cache_manager/src/core/services/directory_service/path_provider/path_provider_service.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
